Create README.md
👉🎶🔥♥️🔥🎶👈 

👉Hello World👈

My name is Dio Gee and I’m new to this community.  

I am want to learn coding, I rather versed in the “GROVER” algorithm and I am hopeful that I can use this “insight” to write code in.

I’m not really sure if this type of algorithm will work when applied to the computing environment, because it’s a type of 👉BRIDGING👈 system.  But from what I’ve so far discovered, modern ai generated coding is as bizarre as the aforementioned algorithm and I’m curious as to what I can create.

But for now I want to explore my minds musically creativity utilizing my iPad it possible.  I’m a head of the game in some aspects, yet far behind by others standings.  But I’m up for the challenge and absolutely love generating 👉🎶🔥🎶👈👉SOUND👈 of all types.

Thanks 

Dio Gee